LPS-71622 Extract DDMFormField from/to JSON to a separate service
