+++ dynamic-data-mapping-io/src/main/java/com/liferay/dynamic/data/mapping/io/internal/DDMFormJSONSerializerImpl.java
@@ -14,29 +14,19 @@
 
 package com.liferay.dynamic.data.mapping.io.internal;
 
-import com.liferay.dynamic.data.mapping.form.field.type.DDMFormFieldType;
-import com.liferay.dynamic.data.mapping.form.field.type.DDMFormFieldTypeServicesTracker;
-import com.liferay.dynamic.data.mapping.form.field.type.DDMFormFieldTypeSettings;
-import com.liferay.dynamic.data.mapping.form.field.type.DefaultDDMFormFieldTypeSettings;
+import com.liferay.dynamic.data.mapping.io.DDMFormFieldJSONObjectTransformer;
 import com.liferay.dynamic.data.mapping.io.DDMFormJSONSerializer;
 import com.liferay.dynamic.data.mapping.model.DDMForm;
 import com.liferay.dynamic.data.mapping.model.DDMFormField;
-import com.liferay.dynamic.data.mapping.model.DDMFormFieldOptions;
-import com.liferay.dynamic.data.mapping.model.DDMFormFieldValidation;
 import com.liferay.dynamic.data.mapping.model.DDMFormRule;
 import com.liferay.dynamic.data.mapping.model.DDMFormSuccessPageSettings;
-import com.liferay.dynamic.data.mapping.model.LocalizedValue;
-import com.liferay.dynamic.data.mapping.util.DDMFormFactory;
 import com.liferay.portal.kernel.json.JSONArray;
 import com.liferay.portal.kernel.json.JSONFactory;
 import com.liferay.portal.kernel.json.JSONObject;
-import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.LocaleUtil;
 
 import java.util.List;
 import java.util.Locale;
-import java.util.Map;
-import java.util.Objects;
 import java.util.Set;
 
 import org.osgi.service.component.annotations.Component;
@@ -50,7 +40,7 @@ public class DDMFormJSONSerializerImpl implements DDMFormJSONSerializer {
 
 	@Override
 	public String serialize(DDMForm ddmForm) {
-		JSONObject jsonObject = _jsonFactory.createJSONObject();
+		JSONObject jsonObject = jsonFactory.createJSONObject();
 
 		addAvailableLanguageIds(jsonObject, ddmForm.getAvailableLocales());
 		addDefaultLanguageId(jsonObject, ddmForm.getDefaultLocale());
@@ -65,7 +55,7 @@ public class DDMFormJSONSerializerImpl implements DDMFormJSONSerializer {
 	protected void addAvailableLanguageIds(
 		JSONObject jsonObject, Set<Locale> availableLocales) {
 
-		JSONArray jsonArray = _jsonFactory.createJSONArray();
+		JSONArray jsonArray = jsonFactory.createJSONArray();
 
 		for (Locale availableLocale : availableLocales) {
 			jsonArray.put(LocaleUtil.toLanguageId(availableLocale));
@@ -87,55 +77,6 @@ public class DDMFormJSONSerializerImpl implements DDMFormJSONSerializer {
 		jsonObject.put("fields", fieldsToJSONArray(ddmFormFields));
 	}
 
-	protected void addNestedFields(
-		JSONObject jsonObject, List<DDMFormField> nestedDDMFormFields) {
-
-		if (nestedDDMFormFields.isEmpty()) {
-			return;
-		}
-
-		jsonObject.put("nestedFields", fieldsToJSONArray(nestedDDMFormFields));
-	}
-
-	protected void addProperties(
-		JSONObject jsonObject, DDMFormField ddmFormField) {
-
-		DDMForm ddmFormFieldTypeSettingsDDMForm =
-			getDDMFormFieldTypeSettingsDDMForm(ddmFormField.getType());
-
-		for (DDMFormField ddmFormFieldTypeSetting :
-				ddmFormFieldTypeSettingsDDMForm.getDDMFormFields()) {
-
-			addProperty(jsonObject, ddmFormField, ddmFormFieldTypeSetting);
-		}
-	}
-
-	protected void addProperty(
-		JSONObject jsonObject, DDMFormField ddmFormField,
-		DDMFormField ddmFormFieldTypeSetting) {
-
-		Object property = ddmFormField.getProperty(
-			ddmFormFieldTypeSetting.getName());
-
-		if (property == null) {
-			return;
-		}
-
-		addProperty(
-			jsonObject, ddmFormFieldTypeSetting.getName(),
-			serializeDDMFormFieldProperty(property, ddmFormFieldTypeSetting));
-	}
-
-	protected void addProperty(
-		JSONObject jsonObject, String propertyName, Object propertyValue) {
-
-		if (propertyValue == null) {
-			return;
-		}
-
-		jsonObject.put(propertyName, propertyValue);
-	}
-
 	protected void addRules(
 		JSONObject jsonObject, List<DDMFormRule> ddmFormRules) {
 
@@ -154,7 +95,7 @@ public class DDMFormJSONSerializerImpl implements DDMFormJSONSerializer {
 	}
 
 	protected JSONArray fieldsToJSONArray(List<DDMFormField> ddmFormFields) {
-		JSONArray jsonArray = _jsonFactory.createJSONArray();
+		JSONArray jsonArray = jsonFactory.createJSONArray();
 
 		for (DDMFormField ddmFormField : ddmFormFields) {
 			jsonArray.put(toJSONObject(ddmFormField));
@@ -163,48 +104,8 @@ public class DDMFormJSONSerializerImpl implements DDMFormJSONSerializer {
 		return jsonArray;
 	}
 
-	protected DDMForm getDDMFormFieldTypeSettingsDDMForm(String type) {
-		DDMFormFieldType ddmFormFieldType =
-			_ddmFormFieldTypeServicesTracker.getDDMFormFieldType(type);
-
-		Class<? extends DDMFormFieldTypeSettings> ddmFormFieldTypeSettings =
-			DefaultDDMFormFieldTypeSettings.class;
-
-		if (ddmFormFieldType != null) {
-			ddmFormFieldTypeSettings =
-				ddmFormFieldType.getDDMFormFieldTypeSettings();
-		}
-
-		return DDMFormFactory.create(ddmFormFieldTypeSettings);
-	}
-
-	protected JSONArray optionsToJSONArray(
-		DDMFormFieldOptions ddmFormFieldOptions) {
-
-		Set<String> optionValues = ddmFormFieldOptions.getOptionsValues();
-
-		if (optionValues.isEmpty()) {
-			return null;
-		}
-
-		JSONArray jsonArray = _jsonFactory.createJSONArray();
-
-		for (String optionValue : optionValues) {
-			JSONObject jsonObject = _jsonFactory.createJSONObject();
-
-			jsonObject.put(
-				"label",
-				toJSONObject(ddmFormFieldOptions.getOptionLabels(optionValue)));
-			jsonObject.put("value", optionValue);
-
-			jsonArray.put(jsonObject);
-		}
-
-		return jsonArray;
-	}
-
 	protected JSONArray ruleActionsToJSONArray(List<String> ruleActions) {
-		JSONArray jsonArray = _jsonFactory.createJSONArray();
+		JSONArray jsonArray = jsonFactory.createJSONArray();
 
 		for (String ruleAction : ruleActions) {
 			jsonArray.put(ruleAction);
@@ -214,7 +115,7 @@ public class DDMFormJSONSerializerImpl implements DDMFormJSONSerializer {
 	}
 
 	protected JSONArray rulesToJSONArray(List<DDMFormRule> ddmFormRules) {
-		JSONArray jsonArray = _jsonFactory.createJSONArray();
+		JSONArray jsonArray = jsonFactory.createJSONArray();
 
 		for (DDMFormRule ddmFormRule : ddmFormRules) {
 			jsonArray.put(toJSONObject(ddmFormRule));
@@ -223,69 +124,12 @@ public class DDMFormJSONSerializerImpl implements DDMFormJSONSerializer {
 		return jsonArray;
 	}
 
-	protected Object serializeDDMFormFieldProperty(
-		Object property, DDMFormField ddmFormFieldTypeSetting) {
-
-		if (ddmFormFieldTypeSetting.isLocalizable()) {
-			return toJSONObject((LocalizedValue)property);
-		}
-
-		String dataType = ddmFormFieldTypeSetting.getDataType();
-
-		if (Objects.equals(dataType, "boolean")) {
-			return GetterUtil.getBoolean(property);
-		}
-		else if (Objects.equals(dataType, "ddm-options")) {
-			return optionsToJSONArray((DDMFormFieldOptions)property);
-		}
-		else if (Objects.equals(dataType, "ddm-validation")) {
-			return toJSONObject((DDMFormFieldValidation)property);
-		}
-		else {
-			return String.valueOf(property);
-		}
-	}
-
-	@Reference(unbind = "-")
-	protected void setDDMFormFieldTypeServicesTracker(
-		DDMFormFieldTypeServicesTracker ddmFormFieldTypeServicesTracker) {
-
-		_ddmFormFieldTypeServicesTracker = ddmFormFieldTypeServicesTracker;
-	}
-
-	@Reference(unbind = "-")
-	protected void setJSONFactory(JSONFactory jsonFactory) {
-		_jsonFactory = jsonFactory;
-	}
-
 	protected JSONObject toJSONObject(DDMFormField ddmFormField) {
-		JSONObject jsonObject = _jsonFactory.createJSONObject();
-
-		addProperties(jsonObject, ddmFormField);
-
-		addNestedFields(jsonObject, ddmFormField.getNestedDDMFormFields());
-
-		return jsonObject;
-	}
-
-	protected JSONObject toJSONObject(
-		DDMFormFieldValidation ddmFormFieldValidation) {
-
-		if (ddmFormFieldValidation == null) {
-			return null;
-		}
-
-		JSONObject jsonObject = _jsonFactory.createJSONObject();
-
-		jsonObject.put(
-			"errorMessage", ddmFormFieldValidation.getErrorMessage());
-		jsonObject.put("expression", ddmFormFieldValidation.getExpression());
-
-		return jsonObject;
+		return ddmFormFieldJSONObjectTransformer.transform(ddmFormField);
 	}
 
 	protected JSONObject toJSONObject(DDMFormRule ddmFormRule) {
-		JSONObject jsonObject = _jsonFactory.createJSONObject();
+		JSONObject jsonObject = jsonFactory.createJSONObject();
 
 		jsonObject.put(
 			"actions", ruleActionsToJSONArray(ddmFormRule.getActions()));
@@ -298,7 +142,7 @@ public class DDMFormJSONSerializerImpl implements DDMFormJSONSerializer {
 	protected JSONObject toJSONObject(
 		DDMFormSuccessPageSettings ddmFormSuccessPageSettings) {
 
-		JSONObject jsonObject = _jsonFactory.createJSONObject();
+		JSONObject jsonObject = jsonFactory.createJSONObject();
 
 		jsonObject.put("body", ddmFormSuccessPageSettings.getBody());
 		jsonObject.put("enabled", ddmFormSuccessPageSettings.isEnabled());
@@ -307,25 +151,11 @@ public class DDMFormJSONSerializerImpl implements DDMFormJSONSerializer {
 		return jsonObject;
 	}
 
-	protected JSONObject toJSONObject(LocalizedValue localizedValue) {
-		Map<Locale, String> values = localizedValue.getValues();
+	@Reference
+	protected DDMFormFieldJSONObjectTransformer
+		ddmFormFieldJSONObjectTransformer;
 
-		if (values.isEmpty()) {
-			return null;
-		}
-
-		JSONObject jsonObject = _jsonFactory.createJSONObject();
-
-		for (Locale availableLocale : localizedValue.getAvailableLocales()) {
-			jsonObject.put(
-				LocaleUtil.toLanguageId(availableLocale),
-				localizedValue.getString(availableLocale));
-		}
-
-		return jsonObject;
-	}
-
-	private DDMFormFieldTypeServicesTracker _ddmFormFieldTypeServicesTracker;
-	private JSONFactory _jsonFactory;
+	@Reference
+	protected JSONFactory jsonFactory;
 
 }
